Reduce Disk I/O usage in ClasspathElementDir
The file attributes are already used when creating a new Resource
so we can easily pass it to the Resource for further usage, thus
can share the already loaded data for a Path.

If from a subPath it turns out to be a file, then we do not need
to check at the end whether it's a directory so we now remove the
files from the pathsInDir.